requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
opencv-python-headless==4.5.5.64
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.1.post1
matplotlib==3.8.3
ipykernel==6.29.0
python-dotenv==1.0.0
Flask-WTF==1.2.1
Pillow==10.3.0
Flask-SQLAlchemy==3.0.5
mysql-connector-python==8.1.0
Flask-Login==0.6.2
email_validator==2.1.0